Fix upgrade process of /etc/casper.conf due to PreBeta ISO snafu
The Jolicloud PreBeta ISO build process manually edited the configuration
file /etc/casper.conf, outside of the casper package itself. In order to
ease upgrades of this package, we must first undo the changes we manually
did, and let dpkg install the correct /etc/casper.conf according to the
package.